Adds capitalization to food button names in index.html
